작업사항(21/12/1~6) - Imgui 연동 중(TileMap, MapTool)
Engine Client

1. GUI
Inspector
GUI_Inspector 클래스 구현 중

GUI_TileMap 클래스 구현

GUI_MapTool 클래스 구현

2. TODO
Hierarchy GUI 버그 수정
Tool용 카메라 구현

Engine Library

1. Component
TileMap 클래스 구현
1) 하나의 TileMap 오브젝트당 최대 8개의 텍스처의 이미지를 넣을 수 있음(같은 셀 사이즈일 경우)

2. ShaderFiles
TileMapShader.fx 구현

3. TODO
Animator2D와 TileMap 파일로 저장하는 기능 추가

+++ External/Include/DX11_2D_GameEngine_Lib/Animation2D.h
@@ -50,7 +50,7 @@ public:
 	const int& GetCurrentFrameID() const { return m_current_frame_id; }
 
 	//현재 애니메이션 프레임에 대한 정보
-	const int& GetAnimationFrameCount() const { return m_p_animation_frame_vector.size(); }
+	const int& GetAnimationFrameCount() const { return static_cast<int>(m_p_animation_frame_vector.size()); }
 	const bool& HasAnimationFrame() const { return !m_p_animation_frame_vector.empty(); }
 	const Animation2D_Frame& GetAnimationFrame(const UINT& frame_index)
 	{
